Убирает токен из localStorage
