+++ project/src/index.js
@@ -1,26 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
-import { composeWithDevTools } from 'redux-devtools-extension';
-import { reducer } from './store/reducer';
-import { createApi } from './services/api';
-import { ActionCreator } from './store/action';
+import { store } from './store/store';
 import { checkAuth, fetchOffers } from './store/api-action';
-import { redirect } from './store/middlewares/redirect';
 import App from './components/app/app';
-
-const api = createApi(
-  () => store.dispatch(ActionCreator.unAuthorize()),
-);
-const store = createStore(
-  reducer,
-  composeWithDevTools(
-    applyMiddleware(thunk.withExtraArgument(api)),
-    applyMiddleware(redirect),
-  ),
-);
 
 store.dispatch(checkAuth());
 store.dispatch(fetchOffers());
