+++ project/src/store/api-action.js
@@ -10,20 +10,21 @@ export const fetchOffers = () => (dispatch, _getState, api) => {
     });
 };
 
-export const checkAuth = () => (dispatch, _getState, api) => (
-  api.get(ApiRoute.LOGIN)
+export const checkAuth = () => (dispatch, getState, api) => {
+  const token = getState().authInfo.userData?.token || '';
+
+  return api.get(ApiRoute.LOGIN, {headers: {'x-token': token}})
     .then(({data}) => dispatch(ActionCreator.authorize({
       status: AuthorizationStatus.AUTH,
       userData: adaptUserDataToClient(data),
     })))
     .then(() => dispatch(ActionCreator.redirect(AppRoute.ROOT)))
-    .catch(() => {})
-);
+    .catch(() => {});
+};
 
 export const login = ({login: email, password}) => (dispatch, _getState, api) => (
   api.post(ApiRoute.LOGIN, {email, password})
     .then(({data}) => {
-      localStorage.setItem('token', data.token);
       dispatch(ActionCreator.authorize({
         status: AuthorizationStatus.AUTH,
         userData: adaptUserDataToClient(data),
@@ -34,7 +35,6 @@ export const login = ({login: email, password}) => (dispatch, _getState, api) =>
 
 export const logout = () => (dispatch, _getState, api) => (
   api.delete(ApiRoute.LOGOUT)
-    .then(() => localStorage.removeItem('token'))
     .then(() => dispatch(ActionCreator.unAuthorize()))
 );
